update filename and add readme
