+++ docs/llm/auto_parallel/README.md
@@ -1,0 +1,1 @@
+../../../llm/auto_parallel/README.md
